final_model/requirements.txt
pandas
numpy
streamlit
pillow
requests
spacy
nltk